+++ apps/src/componentLibrary/alert/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.1.0](https://github.com/code-dot-org/code-dot-org/pull/54064)
+## [0.1.0](https://github.com/code-dot-org/code-dot-org/pull/59159)
 
+* implemented component
+* added tests
+* added storybook
+* component is now ```Ready for Dev```
 * Initial release